refactor(Hero): Move hero parser out of view-matcher and into a helper
