+++ server/views/page-about.html
@@ -1,5 +1,5 @@
 <section class="post hero">
-    {{{header}}}
+    {{{hero header}}}
     <div class="row intro">
         <div class="copy medium-centered medium-8 columns">
             <h2>{{title}}</h2>
